move palettes to their own package. add mono palette
