+++ core/src/com/bitdecay/game/unlock/palette/PaletteList.java
@@ -2,6 +2,7 @@ package com.bitdecay.game.unlock.palette;
 
 import com.bitdecay.game.scoring.ScoreStamps;
 import com.bitdecay.game.system.render.GamePalette;
+import com.bitdecay.game.unlock.palette.types.*;
 
 /**
  * Created by Monday on 6/5/2017.
@@ -11,7 +12,8 @@ public enum PaletteList {
     STANDARD("Helm", new StandardPalette(), 0),
     GUACAMOLE("Guacamole", new GuacamolePalette(), 100000),
     WHITEOUT("Whiteout", new WhiteoutPalette(), 200000),
-    VIDEO_KID("Video Kid", new GameBoyPalette(), 300000);
+    VIDEO_KID("Video Kid", new VideoKidPalette(), 300000),
+    MONOCHROME("Monochrome", new MonochromePalette(), 400000);
 
     public final String name;
     public final int pointsForUnlock;
